Add task wrapper used to update gradle-wrapper to v. 2.2.1 [skip ci]
Add task wrapper to the prject build file.
Change distributionUrl to use https and only binary zip.

The project is currently using gradle-wrapper.jar version 1.6 and downloading
gradle sources each build.

If the task is executed, gradle-wrapper.jar and gradle-wrapper.properties
are updated to this configuration and include changes about gradle user home.
http://www.gradle.org/docs/current/userguide/gradle_wrapper.html

We improve CI avoiding the extra download.
The files gradlew and gradlew.bat are not modified.

There is an alternative solution: execute `./gradlew wrapper` on CI build
when the version is checked but it doesn't help locally.

Next commit completes this pull request, adds the generated gradle-wrapper and
doesn't skip CI build to test it.

I'm worried about introduce binary changes to a public project, last commit can
be reverted or the task executed another time by a Google member (or use the
github/gradle version, it's 2-2-1-rc1).

This is part of a greater CI build fix (currently always fails) and the gradle
user home can be helpful to customize and improve it, add cache option etc.