Maps updates in preparation for new venue.
Replace "Moscone" with "venue" throughout.
Remove use of indoor maps API features. (We only have one floor this year.)
Add support for definition of Maps API key in gradle build config.
Re-implement 'defocus venue' logic to use camera position instead of indoor features.
Update bootstrap data with dummy data for new location.

If the camera is moved away from the venue (for now defined as a
location) or the zoom level is too high, the venue is defocussed
and the large venue marker is displayed.

Change-Id: I154942677556179d782f9d3a94614941f9f4ad17