Add map editor that enables testing of map data.
The map editor is defined in a separate product flavor ("mapEditor")
that replaces the launcher activity and disables fetching of remote data.
All data is read in from a bootstrap file on external storage, every time
the app is opened.
This makes it easy to iterate on the map data very quickly, by simply updating
a json file on the device and restarting the app.

The editor also includes an option to make all markers draggable. Events and
locations are printed to logcat and displayed in the app.

Change-Id: Ic7699ac899695b65b23a1bd88b4609f8bca1ebb7