Removed decline and cancel buttons from welcome flow.
Changed the order of the welcome fragments so that the account selection
fragment precedes the fragment with the in-person/remote options. The first
three fragments present the user with a single, mandatory "Accept" button, and
the fourth fragment presents the users with a choice (and two buttons).

Replaces the idea of "positive" and "negative" buttons with "primary" and
"secondary" buttons (since there is nothing positive or negative about picking
whether you are attending the conference in person or remotely).

Bug: 28075222
Bug: 28139474
Change-Id: Ia27b70fed4d4a028d1dc186fdecb6f48f71d6f3d