Storing both starred and unstarred sessions in local db after sync.
Right now, the sync logic is designed to only store sessions that are in a
user's schedule in the local db. Sessions no longer in the schedule are
deleted. This is changed here.

1) We no longer delete unstarred sessions from the local db.
2) in both Firebase and local db, starred sessions have their in_schedule value
set to true, and unstarred sessions have it set to false (in local db this is
stored as 1 and 0).
3) Starring and unstarring a session updates the timestamp for that session in
Firebase. We compare timestamps to figure out the definitive state of a
session.

Bug: 28325304
Change-Id: I6e5524a0c9a579e227567303ec6d315669703d71