Fix session detail FAB icon anim.
We were overriding the icon in the FAB which was short circuiting
the icon animation. Introduced a check if the checked state is
changing before setting the icon directly to let anim run.

Also removed the state-list drawable setting the icon & always
manually set the correct icon to resolve state issue.

Bug: 28391087
Bug: 28356943
Change-Id: Icc0b4ad73b98f1b1975980a6939073c64fe77563