Interface Android app with Registration API
Adds hook during sign-on to see if the user is registered for the
conference, stores result in shared preferences.

Work is performed in a Service, in order to encapsulate expensive
network calls.

In a future commit, we need to determine how to handle failures.
We could either make login a modal process, blocking until this
is complete. Alternatively, we could record failure and retry at
a later time.

Bug: 36776538
Change-Id: Idfab299d185c55ca2f392ee264edcbebc86f8154