Fix query for sessions in user's schedule
Unfortunately we need to add session_reservation_status to the
projection wherever we query with Sessions.CONTENT_MY_SCHEDULE_URI,
otherwise the generated SQL is malformed. The proper thing to do would
be to use a view in the SQL, but such a change is out of scope for now.

Bug: 37360331
Change-Id: I6f190c13bf117bcca15347377e5245335876e34d