Fix race condition / crash while fetching Firebase token
Fetching a Firebase token is an async operation. We need to wait
for the operation to complete before calling getResult(). Uses a
latch to block the RegistrationStatusService until the token is
available.

Also hardens exception handling, since errors in this service
shouldn't be user visible. (Reg check will be automatically
retried later.)

Bug: 38205069
Change-Id: Ic6b70fb6b19f1722d090e567c7aedd352e0c0615