Add speaker list
For discussion - with a large (20) speaker list, all 20 speaker views
are created in the RecyclerView, negating the benefit of RecyclerView
creating views as needed. Thus, there's no advantage versus creating
a LinearLayout and programatically adding the Views.

We can research using Epoxy, but wanted to discuss with the group
before heading that route.

Conclusion - implementing speaker list in a Linear Layout.

Bug: 72095015
Test: Ran SessionDetailViewModel unit tests

Change-Id: I755be47ed55f5346da5c77cbd9409c9748094a9b