Remove the underscores from the prefix and the suffix for the id for the keynote.
It turned out Firestore doesn't accept a document with id that starts
and ends with "_".
As a result, the document that represents UserEvent for the keynote wasn't stored in Firestore
including its starred state.

Confirmed that the document in Firestore and the star state in the emulator worked
After I manually replaced the "__keynote__" with "keynote".

The copy of the JSON seems to be placed at according to the comment in ConferenceDataDownloader
"https://firebasestorage.googleapis.com/v0/b/iosched-playground.appspot.com/o/" +
        "conference_data.json?alt=media&token=0c0a7ce8-f582-4ab2-a66e-29f335680399"
It needs to be also replaced.

Bug: 75013503
Change-Id: I04a582b42fec0dfddeb91a558d2688ef29d268b6