Work around BottomAppBar shadow issue b/72034805.
The BottomAppBar componenet currently cannot draw an appropriate
elevation shadow for the cradle shape as it cannot draw beyond it's
bounds nor can the platform elevation system draw a concave shaped
shadow.

This workaround introduces a parallel view the same size as the
BottomAppBar with a 9-patch background, faking the shadow.

Change-Id: I6b8436d5f164f097db33d1ec7d85edb86e31bb07