Show livestreamed icon on Session Detail screen only if it is available
Hide the icon by default and show it if the Session's isLivestream
boolean == true.

To test - look at the session detail page for Keynote Boostrap (first
session on day 1) - it should not have a video icon.  The "What's new
in Android" session (May 10, first session) is livestreamed and the
icon should be present in the session detail page.

Most of the changes are updating tests.

Bug: 74617185
Test: kokoro
Change-Id: Iad1f6cbea2a7abbda28588a1678d0d906ea27989