Set the default null value for the observableUserEvents by setting it.
Every time getObservableUserEvents is called, MediatorLiveData.removeSource
and MediatorLiveData.addSource is called in DefaultSessionAndUserEventRepository.
When addSource is called MediatorLiveData considers the old value as new
data coming in (even if the data is for another user).

This CL is to prevent that by setting the null default value by setValue

Bug: 78444538
Change-Id: I3ef96d26e51c166f556f2d3779e123b37efd7006