Update launcher intent-filters
This is an attempt to show the launcher icon
in an instant experience. During development this
didn't change anything, which is expected for an
installed app since both activity and activity-alias
point to the same activity.

When deployed through Play this will show up in the
launcher in an instant experience.

Change-Id: I08d2efe70fd607d047394dd3b5ffd1bb41874b94