Run code style checks on pre-submit
We now use the Spotless plugin to run ktlint and
keep our code style in check. It also does copyright header
checking.

This CL also updates the build script to run the `check`
task which will invoke spotless and lint, to make sure
we don't commit things that shouldn't be.

Change-Id: Ief6510eb8018ec4d40f996fbec3beb7fb6aef005