Initial implementation of dark mode
Currently the UI setting is a binary toggle, but the internal logic is
an enum. This will be moved to a list in a dialog in a later commit.

There are a lot of theme/color issues, so the app looks pretty ugly.
This will also be addressed in a later commit.

Bug: 140107505
Change-Id: I06a387e20c5f0d9bd12fdcd5e300777411eac160