Make app go edge-to-edge
This CL makes the majority of the app draw from top
to bottom, behind the status and navigation bars.

I haven't touch session or speaker details yet,
that will come in a later CL.

The custom build of MDC is because it includes
an upcoming bug fix in AppBarLayout which will
be available in v1.1.0-alpha05.

Change-Id: I61ecf1ea6e17e1868388d18fbdd29f34f1559e8a