Use new WindowInsets APIs on Android Q
Allows us to use the new 'tappable element' insets
to move our views so that they do not conflict with
any system-ui gestures.

Since we can not depend on the Q SDK yet, we need to use
reflection to retrieve the values. This is done in Java
so that we do not need to depend on kotlin-reflect.

BUG: 128876715

Change-Id: I237bff15f92385001813b5c7fa5f3f68db977cd4