Improve Map FAB and bottom sheet interactions
The FAB slides upward to make room when the bottom sheet peeks. If the
sheet starts expanding further than that, the FAB will hide until the
sheet at least reaches the collapsed state.

Since the user can change map variants when the sheet is peeking, and
that particular marker might not be available in the new variant, clear
and hide the sheet if they select another variant.

Also, this makes the sheet content alpha transition more gradual by
using the slide offset instead of the state.

Bug: 129310410
Change-Id: I06e4013318648d61d74b5074b37d6c3e0cec93f8