Update Session Details and sub-navigation structure to use the navigation component.
The session detail and speaker screens are part of the nav graph.
The schedule and feed screens can launch the session details using the nav controller.
The session detail screen can launch other session details (related sessions) and speaker profile pages.
Shared element transitions have been moved from Activity level to Fragment level. Since multiple speakers can appear in session details, the speaker's id became the shared element key. Upon returning from the speaker screen to the session detail screen, if the speaker is not in view, the shared element transition will be skipped.
Handling launching the map from the session details screen is out of scope for this CL.
In a follow up CL, the SessionDetailActivity will be deleted. In the meantime, the SessionDetail fragment assumes that it can be launched from the SessionDetailActivity or from the nav controller.

Bug: 123673505
Bug: 129550095
Change-Id: I465dcfa7de33f89fe8c7128bd49fc33668b6eeef