Tidy up inset and gesture handling for Android Q
Make the Drawer work with system gestures on Q.
Eventually we'll be able to depend on an updated
version of DrawerLayout in AndroidX, but today is
not that day.

Also removed all references to tappable element insets
which are the wrong insets for us to use. System window
insets should be used normally, with system gesture
insets needed for views which scroll (such as bottom
sheets).

Also fixed the logic for closing the drawer from 'back'
gesture. We now disable that action when gesture navigation
is enabled.

BUG: 128876715

Change-Id: I41b341c3f23dbfad201c7857738a364c5d10ed9a