Disable the nav drawer when not in a top-level destination.
This prevents the user from being able to swipe out the nav drawer on secondary screens such as session detail or speaker detail. Also, since the map has an entry in the nav graph for both top level and secondary, depending on how the user gets to the map, the drawer may not be swipable. From the top level, it is swipable but if the user lands on the map from a sub-level, then the drawer is locked closed.

Bug: 130387326
Change-Id: I58994b66f638ee84e4c0e4e414553062a4853b06