Fix the issue announcements are being loaded indefinitely.
Announcements were loaded by subscripting to the change of
the announcement entity in Firestore. But in a following case,
Result.Loading was returned as a result of subscripted announcements.

1) Open Home
   At this moment, the FirestoreAnnouncementDataSource starts to
   subscript to the change of announcement entity.

2) Navigate to another destination

3) Navigate to Home
   At this moment, there was a case where LoadAnnouncementUseCase
   returns Result.Loading and never updated any changes of announcement
   entity because no changes were deteced to the announcement entity
   between 1 and 3.

Given that announcements are almost static except for an explicit push
from spreadsheet, the loading part of announcements can/should be
executed once instead of subscripting to the change.

For the same reason, this CL changes the code to load moments executed
once synchronously.

Bug: 130782854

Change-Id: I9e783194189e273f369b9eba9a04b9fc8789c861