Update Session Detail screen
Moves the photo & video header into a CollapsingToolbarLayout. If an
event has a photo url and/or a video url, we show a full size header
with the photo content (or a placeholder if no photo url) and a play
button if there is a video url.

If an event has neither a photo url nor a video url, we show a "narrow"
header image behind the toolbar based on the event type.

This change also removes a lot of elements:
- The duplicate title that was sometimes shown in the header last year
- The Lottie animations shown in the header last year
- The HeaderGridDrawable
- The PushUpScrollListener behavior that didn't work consistently
- The BottomAppBar shadow drawable, since it is no longer necessary

Bug: 130909132

Change-Id: Ie7a5c28b56f1296de1fe791f3986e137e649b7f7